+++ plugins/input/sqlite/build.py
@@ -1,7 +1,7 @@
 #
 # This file is part of Mapnik (c++ mapping toolkit)
 #
-# Copyright (C) 2007 Artem Pavlenko, Jean-Francois Doyon
+# Copyright (C) 2013 Artem Pavlenko, Jean-Francois Doyon
 #
 # Mapnik is free software; you can redistribute it and/or
 # modify it under the terms of the GNU Lesser General Public
@@ -17,26 +17,24 @@
 # License along with this library; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 #
-# 
+#
 
 Import ('plugin_base')
 Import ('env')
 
-prefix = env['PREFIX']
+PLUGIN_NAME = 'sqlite'
 
 plugin_env = plugin_base.Clone()
 
-sqlite_src = Split(
+plugin_sources = Split(
   """
-  sqlite_datasource.cpp
-  sqlite_featureset.cpp      
-  """
-        )
-
-libraries = [ 'sqlite3' ]
+  %(PLUGIN_NAME)s_datasource.cpp
+  %(PLUGIN_NAME)s_featureset.cpp      
+  """ % locals()
+)
 
 # Link Library to Dependencies
-libraries.append('mapnik')
+libraries = [ 'sqlite3' ]
 libraries.append(env['ICU_LIB_NAME'])
 libraries.append('boost_system%s' % env['BOOST_APPEND'])
 libraries.append('boost_filesystem%s' % env['BOOST_APPEND'])
@@ -45,11 +43,27 @@ linkflags = env['CUSTOM_LDFLAGS']
 if env['SQLITE_LINKFLAGS']:
     linkflags.append(env['SQLITE_LINKFLAGS'])
 
-input_plugin = plugin_env.SharedLibrary('../sqlite', source=sqlite_src, SHLIBPREFIX='', SHLIBSUFFIX='.input', LIBS=libraries, LINKFLAGS=linkflags)
+if env['PLUGIN_LINKING'] == 'shared':
+    libraries.append('mapnik')
 
-# if the plugin links to libmapnik ensure it is built first
-Depends(input_plugin, env.subst('../../../src/%s' % env['MAPNIK_LIB_NAME']))
+    TARGET = plugin_env.SharedLibrary('../%s' % PLUGIN_NAME,
+                                       SHLIBPREFIX='',
+                                       SHLIBSUFFIX='.input',
+                                       source=plugin_sources,
+                                       LIBS=libraries,
+                                       LINKFLAGS=linkflags)
 
-if 'uninstall' not in COMMAND_LINE_TARGETS:
-    env.Install(env['MAPNIK_INPUT_PLUGINS_DEST'], input_plugin)
-    env.Alias('install', env['MAPNIK_INPUT_PLUGINS_DEST'])
+    # if the plugin links to libmapnik ensure it is built first
+    Depends(TARGET, env.subst('../../../src/%s' % env['MAPNIK_LIB_NAME']))
+
+    if 'uninstall' not in COMMAND_LINE_TARGETS:
+        env.Install(env['MAPNIK_INPUT_PLUGINS_DEST'], TARGET)
+        env.Alias('install', env['MAPNIK_INPUT_PLUGINS_DEST'])
+
+plugin_obj = {
+  'LIBS': libraries,
+  'SOURCES': plugin_sources,
+  'LINKFLAGS': linkflags,
+}
+
+Return('plugin_obj')